pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = ""
authors = [
    {name = "jasonli96",email = "jasonli.cv@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pyperclip (>=1.9.0,<2.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "fake-useragent (>=2.1.0,<3.0.0)",
    "stem (>=1.8.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
